requirements.txt
git+https://github.com/successar/allennlp@master#egg=allennlp
torch>=1.2.0
overrides==1.9
numpy==1.16.4
pandas==0.24.2
scripts==2.0
scikit_learn==0.21.3
typing==3.7.4
pandas==0.24.2
Click==7.0
tqdm==4.32.1
fuzzywuzzy==0.17.0
spacy==2.1.6
p_tqdm==1.2
beautifulsoup4==4.8.0
elasticsearch==7.0.2
ipdb
python-intervals
